Provider's data should be in icu_segementer::provider.
- Don't re-export provider's data to icu_segementer
- Move LSTM provider to icu_segmenter::provider
