+++ experimental/segmenter/src/complex.rs
@@ -8,7 +8,6 @@ use icu_provider::{DataError, DataPayload};
 use crate::dictionary::DictionarySegmenter;
 use crate::language::*;
 use crate::provider::*;
-use crate::LstmDataV1Marker;
 
 // Use the LSTM when the feature is enabled.
 #[cfg(feature = "lstm")]
