+++ provider/testdata/data/baked/segmenter/lstm_v1.rs
@@ -1,8 +1,9 @@
 // @generated
-type DataStruct = <::icu_segmenter::LstmDataV1Marker as ::icu_provider::DataMarker>::Yokeable;
+type DataStruct =
+    <::icu_segmenter::provider::LstmDataV1Marker as ::icu_provider::DataMarker>::Yokeable;
 pub static DATA: litemap::LiteMap<&str, &DataStruct, &[(&str, &DataStruct)]> =
     litemap::LiteMap::from_sorted_slice_unchecked(&[("th", TH)]);
-static TH: &DataStruct = &::icu_segmenter::LstmDataV1 {
+static TH: &DataStruct = &::icu_segmenter::provider::LstmDataV1 {
     model: alloc::borrow::Cow::Borrowed("Thai_codepoints_exclusive_model4_heavy"),
     dic: unsafe {
         #[allow(unused_unsafe)]
@@ -69,7 +70,7 @@ static TH: &DataStruct = &::icu_segmenter::LstmDataV1 {
             },
         )
     },
-    mat1: ::icu_segmenter::LstmMatrix {
+    mat1: ::icu_segmenter::provider::LstmMatrix {
         dim: unsafe { ::zerovec::ZeroVec::from_bytes_unchecked(&[74u8, 0u8, 40u8, 0u8]) },
         data: unsafe {
             ::zerovec::ZeroVec::from_bytes_unchecked(&[
@@ -1032,7 +1033,7 @@ static TH: &DataStruct = &::icu_segmenter::LstmDataV1 {
             ])
         },
     },
-    mat2: ::icu_segmenter::LstmMatrix {
+    mat2: ::icu_segmenter::provider::LstmMatrix {
         dim: unsafe { ::zerovec::ZeroVec::from_bytes_unchecked(&[40u8, 0u8, 108u8, 0u8]) },
         data: unsafe {
             ::zerovec::ZeroVec::from_bytes_unchecked(&[
@@ -2437,7 +2438,7 @@ static TH: &DataStruct = &::icu_segmenter::LstmDataV1 {
             ])
         },
     },
-    mat3: ::icu_segmenter::LstmMatrix {
+    mat3: ::icu_segmenter::provider::LstmMatrix {
         dim: unsafe { ::zerovec::ZeroVec::from_bytes_unchecked(&[27u8, 0u8, 108u8, 0u8]) },
         data: unsafe {
             ::zerovec::ZeroVec::from_bytes_unchecked(&[
@@ -3388,7 +3389,7 @@ static TH: &DataStruct = &::icu_segmenter::LstmDataV1 {
             ])
         },
     },
-    mat4: ::icu_segmenter::LstmMatrix {
+    mat4: ::icu_segmenter::provider::LstmMatrix {
         dim: unsafe { ::zerovec::ZeroVec::from_bytes_unchecked(&[108u8, 0u8]) },
         data: unsafe {
             ::zerovec::ZeroVec::from_bytes_unchecked(&[
@@ -3431,7 +3432,7 @@ static TH: &DataStruct = &::icu_segmenter::LstmDataV1 {
             ])
         },
     },
-    mat5: ::icu_segmenter::LstmMatrix {
+    mat5: ::icu_segmenter::provider::LstmMatrix {
         dim: unsafe { ::zerovec::ZeroVec::from_bytes_unchecked(&[40u8, 0u8, 108u8, 0u8]) },
         data: unsafe {
             ::zerovec::ZeroVec::from_bytes_unchecked(&[
@@ -4835,7 +4836,7 @@ static TH: &DataStruct = &::icu_segmenter::LstmDataV1 {
             ])
         },
     },
-    mat6: ::icu_segmenter::LstmMatrix {
+    mat6: ::icu_segmenter::provider::LstmMatrix {
         dim: unsafe { ::zerovec::ZeroVec::from_bytes_unchecked(&[27u8, 0u8, 108u8, 0u8]) },
         data: unsafe {
             ::zerovec::ZeroVec::from_bytes_unchecked(&[
@@ -5786,7 +5787,7 @@ static TH: &DataStruct = &::icu_segmenter::LstmDataV1 {
             ])
         },
     },
-    mat7: ::icu_segmenter::LstmMatrix {
+    mat7: ::icu_segmenter::provider::LstmMatrix {
         dim: unsafe { ::zerovec::ZeroVec::from_bytes_unchecked(&[108u8, 0u8]) },
         data: unsafe {
             ::zerovec::ZeroVec::from_bytes_unchecked(&[
@@ -5829,7 +5830,7 @@ static TH: &DataStruct = &::icu_segmenter::LstmDataV1 {
             ])
         },
     },
-    mat8: ::icu_segmenter::LstmMatrix {
+    mat8: ::icu_segmenter::provider::LstmMatrix {
         dim: unsafe { ::zerovec::ZeroVec::from_bytes_unchecked(&[54u8, 0u8, 4u8, 0u8]) },
         data: unsafe {
             ::zerovec::ZeroVec::from_bytes_unchecked(&[
@@ -5907,7 +5908,7 @@ static TH: &DataStruct = &::icu_segmenter::LstmDataV1 {
             ])
         },
     },
-    mat9: ::icu_segmenter::LstmMatrix {
+    mat9: ::icu_segmenter::provider::LstmMatrix {
         dim: unsafe { ::zerovec::ZeroVec::from_bytes_unchecked(&[4u8, 0u8]) },
         data: unsafe {
             ::zerovec::ZeroVec::from_bytes_unchecked(&[
